% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockEvents.R
\name{getDailyReturnsPerColumn}
\alias{getDailyReturnsPerColumn}
\title{Get the daily returns for each column in a data.frame}
\usage{
getDailyReturnsPerColumn(x)
}
\arguments{
\item{x}{a data.frame of one or more columns to get returns for.}
}
\value{
a data.frame of the stock's daily returns.
}
\description{
Get the daily returns for each column in a data.frame
}

