% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockEvents.R
\name{GetEMACrossOverEvents}
\alias{GetEMACrossOverEvents}
\title{Get events for the EMA crossover technical indicator.}
\usage{
GetEMACrossOverEvents(x, fastEma = 12, slowEma = 26, type = "buy")
}
\arguments{
\item{x}{A single column xts or matrix series with numeric values.}

\item{fastEma}{Number of periods to average over for the faster moving EMA.}

\item{slowEma}{Number of periods to average over for the slower moving EMA.}

\item{type}{buy or sell. Defaults to buy.}
}
\value{
Logical vector where TRUE is the row that the event occured.
}
\description{
Get events for the EMA crossover technical indicator.
}
\examples{
\dontrun{
data(YHOO)
GetEMACrossOverEvents(Ad(YHOO), 12, 26)
}
}

