% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetReturnPercent}
\alias{GetReturnPercent}
\title{Get the return percent from a vector where the first value is the start
and the last value is the end.}
\usage{
GetReturnPercent(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value for the return
}
\description{
Get the return percent from a vector where the first value is the start
and the last value is the end.
}

