% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetValueByDate}
\alias{GetValueByDate}
\title{Get the value of a time series vector by date.}
\usage{
GetValueByDate(x, ind)
}
\arguments{
\item{x}{A time based object.}

\item{ind}{An xts vector.}
}
\value{
A value for that date
}
\description{
Get the value of a time series vector by date.
}

