% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetWindow}
\alias{GetWindow}
\title{Get a window of values starting at an event and ending w instances
after the event.}
\usage{
GetWindow(x, event, w)
}
\arguments{
\item{x}{A vector of values within a time series type object.}
}
\value{
A time series vector of values.
}
\description{
Get a window of values starting at an event and ending w instances
after the event.
}

