% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetWindowCumulativeReturns}
\alias{GetWindowCumulativeReturns}
\title{Get a window of cumulative returns starting at an event and ending w instances
after the event.}
\usage{
GetWindowCumulativeReturns(x, event, w)
}
\arguments{
\item{x}{A numeric vector within a time series type object.}
}
\value{
A time series vector of cumulative returns.
}
\description{
Get a window of cumulative returns starting at an event and ending w instances
after the event.
}

