% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetDailyReturnsPerColumn}
\alias{GetDailyReturnsPerColumn}
\title{Get the daily returns for each column in a data.frame}
\usage{
GetDailyReturnsPerColumn(x)
}
\arguments{
\item{x}{A data.frame of one or more columns to get returns for.}
}
\value{
A data.frame of daily returns for each column.
}
\description{
Get the daily returns for each column in a data.frame
}

