% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetWindowCumulativeReturns}
\alias{GetWindowCumulativeReturns}
\title{Get a window of cumulative returns starting at an event and ending w instances
after the event.}
\usage{
GetWindowCumulativeReturns(x, event, w)
}
\arguments{
\item{x}{An xts vector of numeric values.}

\item{event}{A time based object.}

\item{w}{The number of periods after and including the event.}
}
\value{
A time series vector of cumulative returns.
}
\description{
Get a window of cumulative returns starting at an event and ending w instances
after the event.
}

